[lib] Moved operations common to graphics styles
- Add: Added `GraphicsImage._get_minimal_render_size()`.
  - Moved from `KittyImage._render_image()`.
- Change: Moved methods applicable to multiple graphics style from `KittyImage` into `GraphicsImage`.
  - `_get_render_size()`
  - `_pixels_cols()`
  - `_pixels_lines()`
